Moving everything over to a single function to get lists of tasks
