+++ src/src/styles/index.scss
@@ -1,3 +1,3 @@
 body {
-  background-color: blue;
+  background-color: red;
 }
